Fix a couple bugs in the client processor.
Noticed that when I connected with `nc localhost 8023` then immediately
ctrl-c'd, the server didn't acknowledge the lost connection. Then, when
I tried ctrl-c'ing the server, it would hang waiting for one connection
to reap! Both of these bugs seem to be caused by the same (lack of)
pattern in `tokio::select!()`:

1. `Some(line) = lines.next` instead of `opt_line = lines.next()`. Turns
out that the contract of streams is that they return `None` when finished,
and we weren't catching that case. This prevented the connection task from
completing when the connection dropped.

2. `Some(true) = stopped_rx.next()` instead of `Some(stop) = stopped_rx.next`.
This prevented the connection task from exiting even when the main task
trigged the watch.