Refactor `server::run` to be testable; add a smoke test for it.
The function now takes a `oneshot::Receiver` for shutdown notifications,
and its loop selects on that instead of `tokio::signal::ctrl_c`. This
makes it easier to write a test for proper shutdown behavior. To do this,
`main` spawns a task for `server::run`, and awaits on the signal itself.
Once received, it sends the notice through the oneshot channel.

Writing a test for this led to some interesting findings. The server kept
reporting fewer accepted connections that the number of clients, even
after making each client write nonsense. By having each client send a
valid request and wait for a response, however, the numbers matched.