Decouple connection processing and request handling into two tasks.
Assuming three players, A, B, and C, are in the same room, and requests
from all three are received around roughly the same time. One processing
task, let's say A's, wins the race for the mutex lock, and executes it.
This likely will generate unsolicited responses to be send to B and C.
Neither, however, are receiving the unsolicited responses, since their
respective singular processing tasks are busy waiting to acquire the lock.

By separating (de)serialization of requests and responses from the TCP
stream and handling of requests, this can be avoided. Only the request
handling tasks for A, B, and C are fighting for the lock, while the ones
processing the connections are ready.

This commit also changes the to use bounded mpsc channels, so that there's
back-pressure and a limit on the amount of pending work that can
accumulate.