Add containerized build option.
Currently, this just builds a package for Debian Buster, but this build
script should be fairly straightforward to make portable to target
multiple Debian-based distros and releases.

Note that we don't build the package within there `Dockerfile`, but instead
define a builder image there which contains all the necessary tooling.
Then, we can run `make pkg` in a container using that image, and mount
the output and Cargo registry directories from the host. This lets rebuilds
complete without needing to re-download and re-compile all dependencies,
while maintaining a good degree of isolation in the build.