Overhaul structure for graceful shutdowns.
While experimenting with the new `smol` executor, I wound up re-writing
the shutdown logic using the MPMC channel from `piper`. Further, the
code is now more idiomatic, and utilizing the property that when all
sending endpoints of a channel are dropped, each receiving endpoint waiting
for a value will get `None`. Instead of a `oneshot` channel notifying the
main server task to shutdown, then a `watch` fanning the notice out to each
client task, instead, everyone can simply hold onto clones of the receiving
endpoint of a single channel.

Even better, we can use a second channel to correctly wait until all client
tasks have terminated. Each client task has a sending endpoint. The main
server task waits for the server to stop, then waits for the termination
receiver. This should ensure that all client tasks have time to clean up
(e.g., send a message to the client that the server is shutting down.)

I also found a peculiar bug while implementing this. Since the `Guard` is
moved into the closure passed to Warp's `map`, and contains a clone of
the termination sender, if it doesn't get dropped, the termination receiver
would never stop waiting. I'm not sure _why_ the closure wasn't getting
cleaned up; it may be a bug in Warp, or a bug in my understanding of how
the Warp server lifecycle is defined. To get around this, we instead move
a weak reference to the guard into that closure, and keep a strong
reference in the main server task. The closure will attempt to upgrade the
weak reference when executed, panic if it doesn't succeed, and then drop
the strong reference ASAP.

Further, there seems to be some _other_ bug in Warp that causes the server
never to finish, even when the graceful termination future completes. To
get around this, we launch another task to run and wait on the server, and
don't bother joining on it. If the HTTP server can't stop in time, too bad.

Lastly, before, we assumed that the shutdown signal would always complete
before the server task. If there's a bug, though, and the server task
panics, the main thread will keep hanging awaiting for the signal. Instead,
we should wait for either the signal or the server task to complete.